integration of careers landing page with JD page and form page and Get requests from backend for all jobs and their JDs
